Add monomial Option Prism
