+++ build.sbt
@@ -1,3 +1,4 @@
+val CatsV = "2.10.0"
 val TypenameV = "0.2.0"
 val TypesizeV = "0.1.0"
 
@@ -31,6 +32,9 @@ lazy val poly = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("modules/poly"))
   .settings(
     name := "polynomial",
+    libraryDependencies ++= Seq(
+      "org.typelevel" %%% "cats-core" % CatsV,
+    )
   )
 
 lazy val mermaid = crossProject(JSPlatform, JVMPlatform, NativePlatform)
